# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAphoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can upload your photos directly from your PC, and they will be visible to everyone following you.

    ### Functional design specifications
    Each user will be presented with a stream of photos (images) in reverse chronological order, with information about when each photo was uploaded (date and time) and how many likes and comments it has. The stream is composed by photos from “following” (other users that the user follows).

    Users can place (and later remove) a "like" to photos from other users. Also, users can add comments to any image (even those uploaded by themself). Only authors can remove their comments.

    Users can ban other users. If user Alice bans user Eve, Eve won't be able to see any information about Alice. Alice can decIde to remove the ban at any moment.

    Users will have their profiles. The personal profile page for the user shows: the user's photos (in reverse chronological order), how many photos have been uploaded, and the user's followers and following.

    Users can change their usernames, upload photos, remove photos, and follow/unfollow other users. Removal of an image will also remove likes and comments.

    A user can search other user profiles via username.

    A user can log in just by specifying the username. See the "Simplified login" section for details.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  responses:
    NoContent: #204
      description: The request was successful, but the client doesn't need to navigate away from its current page.
    BadRequest: #400
      description: The request was not compliant with the documentation (eg. missing fields, etc).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized: #401
      description: The request was not completed due to lack of authentication credentials for the requested resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound: #404
      description: The specified resource was not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalServerError: #500
      description: The server encountered an internal error. Further info in server logs.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      title: Error
      description: |-
        This object represents a single error to be returned for certain requests
      type: object
      properties:
        code:
          type: string
          description: An HTTP error code
          example: "404"
        message:
          type: string
          description: A brief message describing the error code
          example: The specified resource was not found
      required: [ "code", "message" ]
...
security:
  - bearerAuth: [ ]
...
paths:
...
  /users/{username}/banned/{bannedUser}:
    parameters:
        - name: username
          in: path
          description: The username of the user who wants to unban someone
          required: true
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User/properties/username'}
        - name: bannedUser
          in: path
          description: The username of the user to be unbanned
          required: true
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User/properties/username'}
    delete:
      tags: [ "theirProfile" ]
      operationId: unbanUser
      summary: unban a user
      description: "Given the usernameof a user, it allows you to unban a certain user"
      responses:
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...
  /users/{username}/posts/{postId}/likes/{likeId}:
    delete:
      tags: [ "Post" ]
      description: "Remove a like from a post"
      operationId: unlikePhoto
      summary: remove like to a post
      parameters:
        - name: postId
          in: path
          description: the Id of the post being unliked
          required: true
          schema: { $ref: '#/components/schemas/Post/properties/postId' }
        - name: username
          in: path
          description: the username of the user whose post is being shown
          required: true
          schema: { $ref: '#/components/schemas/User/properties/username' }
        - name: likeId
          in: path
          description: the Id of the like to remove
          required: true
          schema: {$ref: "#/components/schemas/Like/properties/likeId"}
      responses:
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...
  /users/{username}/posts/{postId}/comments/{commentId}:
    delete:
      tags: ["Post"]
      operationId: uncommentPhoto
      summary: delete your comment on a post
      description: allows you to remove your own comment on a post
      parameters:
        - name: commentId
          in: path
          description: the unique ID of a comment
          required: true
          schema: {$ref: '#/components/schemas/Comment/properties/commentId'}
        - name: username
          in: path
          description: the username of the user whose post is being commented on
          required: true
          schema: {$ref: '#/components/schemas/User/properties/username'}
        - name: postId
          in: path
          description: the unique ID of the post where is the comment
          required: true
          schema: {$ref: '#/components/schemas/Post/properties/postId'}
      responses:
        "204": { $ref: '#/components/responses/NoContent' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "500": { $ref: '#/components/responses/InternalServerError' }
...